requirements.txt
python-telegram-bot==13.15
tensorflow-cpu
numpy
pillow
